Fixed PR fail issue
